requirements.txt
requests
colorama
pycryptodome
bs4